requirements.txt
datadog==0.44.0
pyyaml==6.0
dacite==1.6.0